🚚 [Modules] Move some files to Module/ShaderBased
